@@ -9,11 +9,13 @@
         "request-id"
     ],
     "require": {
-        "php": "^5.5 || ^7.0",
-        "psr/http-message": "^1.0"
+        "php": "^5.6 || ^7.0",
+        "psr/http-message": "^1.0",
+        "php-middleware/double-pass-compatibility": "^1.0",
+        "http-interop/http-middleware": "^0.4.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8.6",
+        "phpunit/phpunit": "^5.6 || ^6.1.3",
         "ramsey/uuid": "^3.0",
         "zendframework/zend-diactoros": "^1.1.3"
     },
