@@ -4,15 +4,15 @@ namespace PhpMiddlewareTestTest\RequestId\Generator;
 
 use PhpMiddleware\RequestId\Generator\GeneratorInterface;
 use PhpMiddleware\RequestId\Generator\Md5Generator;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class Md5GeneratorTest extends PHPUnit_Framework_TestCase
+class Md5GeneratorTest extends TestCase
 {
     protected $generator;
 
     protected function setUp()
     {
-        $decoratedGenerator = $this->getMock(GeneratorInterface::class);
+        $decoratedGenerator = $this->createMock(GeneratorInterface::class);
         $decoratedGenerator->method('generateRequestId')->willReturn('boo');
 
         $this->generator = new Md5Generator($decoratedGenerator);
