@@ -4,15 +4,15 @@ namespace PhpMiddlewareTestTest\RequestId\Generator;
 
 use PhpMiddleware\RequestId\Generator\GeneratorInterface;
 use PhpMiddleware\RequestId\Generator\PrefixedGenerator;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class PrefixedGeneratorTest extends PHPUnit_Framework_TestCase
+class PrefixedGeneratorTest extends TestCase
 {
     protected $decoratedGenerator;
 
     protected function setUp()
     {
-        $this->decoratedGenerator = $this->getMock(GeneratorInterface::class);
+        $this->decoratedGenerator = $this->createMock(GeneratorInterface::class);
         $this->decoratedGenerator->method('generateRequestId')->willReturn('boo');
     }
 
