@@ -3,11 +3,11 @@
 namespace PhpMiddlewareTestTest\RequestId\Generator;
 
 use PhpMiddleware\RequestId\Generator\RamseyUuid4StaticGenerator;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use Ramsey\Uuid\Uuid;
 use Ramsey\Uuid\UuidInterface;
 
-class RamseyUuid4StaticGeneratorTest extends PHPUnit_Framework_TestCase
+class RamseyUuid4StaticGeneratorTest extends TestCase
 {
     protected $generator;
 
