@@ -6,19 +6,17 @@ use PhpMiddleware\RequestId\Exception\InvalidRequestId;
 use PhpMiddleware\RequestId\Exception\MissingRequestId;
 use PhpMiddleware\RequestId\Generator\GeneratorInterface;
 use PhpMiddleware\RequestId\OverridePolicy\OverridePolicyInterface;
-use PhpMiddleware\RequestId\RequestIdMiddleware;
 use PhpMiddleware\RequestId\RequestIdProvider;
-use PHPUnit_Framework_TestCase;
-use Zend\Diactoros\Response;
+use PHPUnit\Framework\TestCase;
 use Zend\Diactoros\ServerRequest;
 
-class RequestIdProviderTest extends \PHPUnit_Framework_TestCase
+class RequestIdProviderTest extends TestCase
 {
     protected $generator;
 
     protected function setUp()
     {
-        $this->generator = $this->getMock(GeneratorInterface::class);
+        $this->generator = $this->createMock(GeneratorInterface::class);
     }
 
     public function testGenerateIdBecauseNotExistInHeader()
@@ -37,12 +35,12 @@ class RequestIdProviderTest extends \PHPUnit_Framework_TestCase
      */
     public function testTryToGenerateEmptyRequestId($emptyValue)
     {
-        $this->setExpectedException(InvalidRequestId::class);
-
         $this->generator->expects($this->once())->method('generateRequestId')->willReturn($emptyValue);
 
         $request = new ServerRequest();
         $provider = new RequestIdProvider($request, $this->generator);
+
+        $this->expectException(InvalidRequestId::class);
 
         $provider->getRequestId();
     }
@@ -57,12 +55,12 @@ class RequestIdProviderTest extends \PHPUnit_Framework_TestCase
 
     public function testTryToGenerateNotStringRequestId()
     {
-        $this->setExpectedException(InvalidRequestId::class);
-
         $this->generator->expects($this->once())->method('generateRequestId')->willReturn(1);
 
         $request = new ServerRequest();
         $provider = new RequestIdProvider($request, $this->generator);
+
+        $this->expectException(InvalidRequestId::class);
 
         $provider->getRequestId();
     }
@@ -91,11 +89,13 @@ class RequestIdProviderTest extends \PHPUnit_Framework_TestCase
 
     public function testDoNotGenerateBecauseHeaderExistsButEmpty()
     {
-        $this->setExpectedException(MissingRequestId::class);
         $this->generator->expects($this->never())->method('generateRequestId');
         $request = new ServerRequest([], [], 'https://github.com/php-middleware/request-id', 'GET', 'php://input', [RequestIdProvider::DEFAULT_REQUEST_HEADER => '']);
 
         $provider = new RequestIdProvider($request, $this->generator);
+
+        $this->expectException(MissingRequestId::class);
+
         $provider->getRequestId();
     }
 
@@ -105,7 +105,7 @@ class RequestIdProviderTest extends \PHPUnit_Framework_TestCase
 
         $request = new ServerRequest([], [], 'https://github.com/php-middleware/request-id', 'GET', 'php://input', [RequestIdProvider::DEFAULT_REQUEST_HEADER => '987654321']);
 
-        $policy = $this->getMock(OverridePolicyInterface::class);
+        $policy = $this->createMock(OverridePolicyInterface::class);
         $policy->method('isAllowToOverride')->with($request)->willReturn(true);
 
         $provider = new RequestIdProvider($request, $this->generator, $policy);
@@ -119,7 +119,7 @@ class RequestIdProviderTest extends \PHPUnit_Framework_TestCase
         $this->generator->method('generateRequestId')->willReturn('123456789');
         $request = new ServerRequest([], [], 'https://github.com/php-middleware/request-id', 'GET', 'php://input', [RequestIdProvider::DEFAULT_REQUEST_HEADER => '987654321']);
 
-        $policy = $this->getMock(OverridePolicyInterface::class);
+        $policy = $this->createMock(OverridePolicyInterface::class);
         $policy->method('isAllowToOverride')->with($request)->willReturn(false);
 
         $provider = new RequestIdProvider($request, $this->generator, $policy);
