@@ -4,9 +4,10 @@ namespace PhpMiddlewareTestTest\RequestId;
 
 use PhpMiddleware\RequestId\RequestDecorator;
 use PhpMiddleware\RequestId\RequestIdProviderInterface;
+use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\RequestInterface;
 
-class RequestDecoratorTest extends \PHPUnit_Framework_TestCase
+class RequestDecoratorTest extends TestCase
 {
     const CUSTOM_HEADER_NAME = 'custom-header-name';
 
@@ -14,14 +15,14 @@ class RequestDecoratorTest extends \PHPUnit_Framework_TestCase
 
     protected function setUp()
     {
-        $requestIdProvider = $this->getMock(RequestIdProviderInterface::class);
+        $requestIdProvider = $this->createMock(RequestIdProviderInterface::class);
         $requestIdProvider->expects($this->once())->method('getRequestId')->willReturn('boo');
         $this->decorator = new RequestDecorator($requestIdProvider, self::CUSTOM_HEADER_NAME);
     }
 
     public function testIsRequestDecorated()
     {
-        $request = $this->getMock(RequestInterface::class);
+        $request = $this->createMock(RequestInterface::class);
         $request->expects($this->once())->method('withHeader')->willReturnCallback(function($name, $value) use ($request) {
             $this->assertSame(self::CUSTOM_HEADER_NAME, $name);
             $this->assertSame('boo', $value);
